feat: chartiq lazy loading
