+++ ui/portal/web/src/pages/__root.tsx
@@ -48,7 +48,12 @@ export const Route = createRootRouteWithContext<RouterContext>()({
       })();
     }, []);
 
-    if (!isReady) return <Spinner />;
+    if (!isReady)
+      return (
+        <div className="flex h-screen w-screen items-center justify-center">
+          <Spinner size="md" color="pink" />
+        </div>
+      );
 
     return (
       <>
